requirements.txt
Flask==2.3.3
Flask-Login==0.6.3
Werkzeug==2.3.7
Jinja2==3.1.2
MarkupSafe==2.1.3
itsdangerous==2.1.2
click==8.1.7
blinker==1.6.3
bleach==6.1.0
python-slugify==8.0.1
Flask-Migrate==4.0.5
Alembic==1.12.0
SQLAlchemy==2.0.21
Flask-Mail==0.9.1
Flask-Caching==2.1.0
Flask-Limiter==3.5.0
gunicorn==21.2.0
psycopg2-binary==2.9.7
Flask-SQLAlchemy==3.0.5
email-validator==2.0.0
Pillow==10.0.1
python-dotenv==1.0.0
WTForms==3.0.1
Markdown==3.5.1
redis==4.6.0

